Move more form-methods into form-mixin + Rename specialProperties to hiddenProperties
